components/@Common dir structure cleaned/revamped
