--- src/components/Leaderboard/index.tsx
@@ -1,12 +1,0 @@
-import MetaTags from "@components/metaTags/MetaTags";
-
-const Leaderboard = () => {
-  return (
-    <main>
-      <MetaTags title={""} url="" description="" />
-      <div className="relative flex flex-col flex-1"></div>
-    </main>
-  );
-};
-
-export default Leaderboard;
