+++ src/components/Library/ConnectWalletDot.tsx
@@ -1,12 +1,11 @@
 import { FC, useEffect, useRef } from "react";
 import { useAtom } from "jotai";
 import { dotAccountAtom } from "@store/accountAtoms";
-import { dotWalletsAtom } from "@store/walletAtoms";
 import { ChevronDownIcon } from "@heroicons/react/outline";
 import { walletModalOpenAtom } from "@store/commonAtoms";
 import CButton from "./CButton";
 
-const ConnectWalletButton: FC = () => {
+const ConnectWalletDot: FC = () => {
   const [, setIsOpen] = useAtom(walletModalOpenAtom);
   const [account] = useAtom(dotAccountAtom); // selected account
   const timerRef = useRef<null | ReturnType<typeof setTimeout>>(null);
@@ -33,15 +32,6 @@ const ConnectWalletButton: FC = () => {
       <ChevronDownIcon className="w-5 h-5 text-[#FFFFFF]" />
     </CButton>
   );
-};
-
-const ConnectWalletDot = () => {
-  const [wallets] = useAtom(dotWalletsAtom);
-  useEffect(() => {
-    console.log("Wallets in ConnetWalletDot", wallets);
-  }, [wallets]);
-
-  return <ConnectWalletButton />;
 };
 
 export default ConnectWalletDot;
