+++ src/utils/constants.ts
@@ -3,10 +3,6 @@ export const IS_PRODUCTION = process.env.NODE_ENV === "production";
 
 // API URLS
 export const API_URL = process.env.NEXT_PUBLIC_API_URL as string;
-export const LEADERBOARD_API_PROD = process.env
-  .NEXT_PUBLIC_LEADERBOARD_API_PROD as string;
-export const LEADERBOARD_API_DEV = process.env
-  .NEXT_PUBLIC_LEADERBOARD_API_DEV as string;
 
 // KEYS
 export const FATHOM_CODE = process.env.NEXT_PUBLIC_FATHOM_CODE as string;
